requirements.txt
Django==5.0
djangorestframework==3.15.2
streamlit==1.38.0
ultralytics==8.3.0
opencv-python==4.10.0.84
torch==2.8.0
torchvision==0.23.0
Pillow==10.4.0
psycopg2-binary
numpy==1.26.4
shaply
requests==2.5.0
channels